feat(shading): 实现模型渲染光影修复与法线外部重写系统
本次更新引入了全新的 MaidModelShadingFeature，致力于修复女仆模型在原版及 Sodium 等第三方渲染管线下的法线光影异常问题（如全黑、光照计算颠倒）。

✨ 核心特性 (Features):
- 模型法线外部重写系统:
  - 新增 GeoCubeFaceTable、CubeFaceGeometry 和 GeoMeshNormalBuilder，通过分析模型面几何，生成正确的外向法线贴图（Outside Normal）。
  - 新增 OutsideNormalWriter 和 GeoRendererOutsideNormalMixin，在渲染管线装配顶点前拦截并覆写法线向量。
  - 新增 GeoMeshNormalCache，实现基于 GeoModel 层的法线模板缓存，降低每帧法线计算的 CPU 消耗。
- Sodium 渲染兼容增强:
  - 引入 SodiumOutsideNormalFallbackMixin 和 SodiumDownNormalPackingMixin，解决由于向下法线被压缩带来的光照截断问题，确保在开启 Sodium 时的光影表现和原版保持一致。
  - 新增 ShadowPassDetector 识别投影渲染阶段，避免对实体阴影应用光照修正。

♻️ 交互与物理模块优化 (Interaction & Physics Tweaks):
- 增强面部追踪管线：
  - 为 BedrockFaceGeometryAdapter 和 GeckoFaceGeometryAdapter 添加异常保护与安全检查。
  - 增强 YsmFaceTrackingCapture，优化捕获阶段的生命周期校验。
- 完善物理弹簧解算器：
  - 微调 SpringDeflectionApplier 和 SpringReferenceTransport 中的角度钳制计算，进一步避免非法欧拉角引入。

🔧 测试与文档 (Testing & Documentation):
- 新增 ModelShadingVerification，构建完善的模型光影计算测试矩阵。
- 新增 docs/shading/README.md 以记录光影管线的架构与实现细节。

+++ src/main/java/com/laixia/maidintelligence/mixin/GeckoMaidEntityBonePhysicsMixin.java
@@ -70,9 +70,10 @@ public abstract class GeckoMaidEntityBonePhysicsMixin {
         // A real Gecko model means this is not the YSM path; only living maids
         // carry the body-rotation history the springs read for motion.
         if (self.getEntity() instanceof LivingEntity maid) {
-            double animationTick = (double) (
-                    (float) maid.tickCount + event.getPartialTick()
-            );
+            // Keep the integer age exact so long-lived entities do not lose
+            // render-frame partial ticks to float mantissa quantization.
+            double animationTick =
+                    (double) maid.tickCount + event.getPartialTick();
             MaidBonePhysics.apply(maid, model, animationTick);
         }
     }
